Service Improvements
- Purchase Fleet Ack Screen should allow XML imports
- If a fleet vehicle is picked on the service order it will now set the contact to that of the fleet
- Adjusted the record rules setup for projects/tasks to allow related companies to view the related projects
- Added is_wip_stage to project tasks. When a task timer is started it will auto move the stage
- Service writer is no longer defaulted to the user logged when autogenerated from an SO/PO
- If the start/end dates are adjusted on a repair order it will now change the tasks dates if they are not in wip/done
- Added constraints to the project task that will block it from picking date ranges outside of the repair order
- Followers/Activities were added to the service order
- Service Templates now require 'type' to prevent downstream issues
- Added exception handling improvements for the fast api sales order route

+++ cap_service/models/service_order_line_service.py
@@ -12,7 +12,7 @@ class ServiceOrderLineService(models.Model):
     qty_to_invoice = fields.Float(compute='_compute_qty_invoice', store=True, readonly=False)
     qty_invoiced = fields.Float(compute='_compute_qty_invoice', store=True, readonly=False)
     product_id = fields.Many2one('product.product')
-    quantity = fields.Integer()
+    quantity = fields.Float()
     unit_price = fields.Float()
 
     def _compute_display_name(self):
